refactor: rename dirs
