+++ src/modules/artists/artist.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { v4 as createId } from 'uuid';
 
-import { Message } from './constants/artist.constants';
+import { Message } from './constants/message.constants';
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
 import { Artist } from './interfaces/artist.interface';
