+++ src/modules/favorites/favorite.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 
-import { AlbumModule } from '../album/album.module';
-import { ArtistModule } from '../artist/artist.module';
+import { AlbumModule } from '../albums/album.module';
+import { ArtistModule } from '../artists/artist.module';
 import { FavoriteService } from './favorite.service';
 import { FavoriteController } from './favorite.controller';
-import { TrackModule } from '../track/track.module';
+import { TrackModule } from '../tracks/track.module';
 
 @Module({
   imports: [AlbumModule, ArtistModule, TrackModule],
