+++ src/modules/favorites/interfaces/favorite.interface.ts
@@ -1,6 +1,6 @@
-import { Album } from '../../album/interfaces/album.interface';
-import { Artist } from '../../artist/interfaces/artist.interface';
-import { Track } from '../../track/interfaces/track.interface';
+import { Album } from '../../albums/interfaces/album.interface';
+import { Artist } from '../../artists/interfaces/artist.interface';
+import { Track } from '../../tracks/interfaces/track.interface';
 
 export interface Favorites {
   artists: string[]; // favorite artists ids
